requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.25
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
Werkzeug==2.3.7
alembic==1.12.1
psycopg2-binary==2.9.9
